Get basic md5 authentication working
This commit gets basic md5 working. The code is pretty straight line
and won't scale well. Parsing knowledge needs to be moved out which
is a refactoring that will be coming soon.

There's a bunch of TODOs that have been generated from this as well
for unit tests and other additional changes that came from getting
the basics working.