Extract standalone response parser (#6)
Previously, all the parsing logic for extracting messages from a buffer
of received data was in the PgSession. I've moved that logic out into
its own parser than can be much more easily tested. The initial tests
will be coming in the next commit.

This commit starts to introduce types for each response type so the
parser can be input of what we do in response to a given message from
the server.