Add basic state machine mechanism to PgSession (#8)
The state machine is currently using a variable to store where we are in
the state machine. This should probably be using the state pattern, but
this is a good first step for getting the mechanism in place.

As part of adding the state machine, we have updated to shutdown should
we encounter an unrecoverable parsing error.