Rename PgSession
The "Pg" is redundant given that our package is "postgres" and if
the user wants to quality names they can do:

```pony
use pg = "postgres"
```

All classes/variables that had Pg as a prefix has been renamed. However,
the "pg_" at the beginning of methods on the session status notify remain.
I left those as it should be possible to construct an actor that receives
session notifications from multiple sources and so, the "pg_" qualifier
feels like it makes sense.