Add early exit in unexpected situations
If something happens that we do not expect to happen, then we will
potentially exit early depending on how unlikely the thing is.

For now, _IllegalState exists are always on. Before a version 1.0.0,
we should make them be only when running debug code.

_Unreachable should always exit as something very very wrong is going
on if we hit code that we believe to be unreachable.