Use state pattern to model PgSession state (#10)
With just a small bit of functionality in place, the state transitions
in PgSession were fairly clear but you could see the outline of trouble
coming.

As an example of "trouble", have a look at TCPConnection in the standard
library. Pulling apart the different bits of the state machine in
TCPConnection and what is and isn't available for any given state is
difficult.

By moving PgSession to the state pattern early, it makes it easy to see
what our various state transitions are and what is and isn't possible at
each point. During development, I plan on putting "let it crash" exits
at any point in the state machine where we encounter unexpected code
being called. The state pattern makes these much easier to find as they
are very explicit. And those "let it crash" exits will make it easier to
find bugs in the code as we develop it.

I've done a couple of interesting things as part of this change. First,
all the state handlers are themselves stateless, leaving all state
storage to the enclosing PgSession actor. All the state fields in the
actor that the state handlers need to access have been made public. By
being public we can access them from our state handlers the same as we
would from inside the actor. There's no need to add a mass of getters
and setters in this case. We simply access the fields as needed. This
isn't problematic from an encapsulation standpoint as objects "external
to" a PgSession actor can't read it's fields even if they are public.
Only objects that we give PgSession ref aliases can access the fields.

We have an interface defined for state handlers, _PgSessionState. The
reason we are using an interface rather than a trait comes down to ease
of constructing handlers.

We have a series of traits for different bits of state that can change.
For example, connected vs not connected, authenticated vs not
authenticated, connectable vs not connectable. Each high level state
mixes some combination of them together.

The functionality for each is in one or more default methods on the
given traits. We need an interface for the overall _PgSessionState so we
can have compile time warnings if we try to assign a state handler that
hasn't been fully assembled (aka is missing methods). We could use a
trait but that then requires an extra `is _PgSessionState`. In the
current context, we get little from trait as we will already get
compilation errors if any state is missing methods when we try to assign
to the `state: _PgSessionState` in the PgSession actor.

I think this has worked out fairly well as we can now easily see our
current states:

```pony
primitive _PgSessionUnopened is _ConnectableState
primitive _PgSessionClosed is (_NotConnectableState & _UnconnectedState)
primitive _PgSessionConnected is _AuthenticableState
primitive _PgSessionLoggedIn is (_ConnectedState & _NotAuthenticableState)
```

It would be nice if the states could be made type-safe and to be compile
time errors if we try to call a method that is "illegal" for the current
state. "Sadly", that isn't possible given the actor based nature of
Pony.

An actor can receive a messge from any other actor at any time. There's
no way at compile time to prevent an actor sending a message of type X
to another actor whose state deems the message "illegal". Further, at
the time that a message is sent, it might be "legal" but by the time it
is processed, it could be "illegal". Such is the way with asychronous
programming. So, we are doing runtime checking. At the moment the
"illegal states" are marked by comments in the code that are there as
placeholders for me to do something more there. See above for ideas on
adding "let it crash" exits.