Add handling all authentication failure types (#14)
Previously, we were only notifying and correctly handling invalid
passwords. If we got an "invalid authentication specification" error,
so say, sending MD5 passwords when they aren't supported, we were
silently ignoring.

We now handle both possible authentication errors. As part of this
change, the session status notify interface for authentication failure
takes a second parameter that indicates the kind of failure.

In the Postgres docs, all messages that we get for authentication
are called authentication messages, but the errors use the terminology
"authorization". I've gone with "authentication" throughout this
codebase.