Enhance shutdown handling
This commit adds a test that when junk messages are received by a session
that it responds by shutting down. In order to write this test, a new
SessionStatusNotify method `pg_session_shutdown` was added.

Without the new method, we would have been unable to write the test and
more importantly, the test exposed that getting session ended notifications
is a key part of the status API that was missing until this commit.