Add basic "Simple Query" support (#27)
There's a lot that still needs to be done for this, however, there's
enough here for basic queries. There's also enough to start getting some
real usage to get feedback on the API.

When Extended Queries are added next, the API will change and a number
of open TO-DOs that haven't been addressed will be addressed as the
Extended Queries work changes the implementation.